backend/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
CodeConvert==3.0.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.8
DAWG-Python==0.7.2
defusedxml==0.8.0rc2
Django==3.2.16
django-admin==2.0.2
django-colorfield==0.11.0
django-excel-response2==3.0.6
django-filter==23.5
django-models-ext==1.1.11
django-six==1.0.5
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
docopt==0.6.2
drf-extra-fields==3.7.0
excel-base==1.0.4
filetype==1.2.0
furl==2.1.3
idna==3.7
isoweek==1.3.3
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
oauthlib==3.2.2
orderedmultidict==1.0.1
pillow==10.4.0
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
pymorphy2==0.9.1
pymorphy2-dicts==2.4.393442.3710985
pymorphy2-dicts-ru==2.4.417127.4579844
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
screen==1.0.1
shortuuid==1.0.13
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.5.4
sqlparse==0.5.0
TimeConvert==3.0.13
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
xlwt==1.3.0
